Currencies

+++ database/migrations/2018_04_17_113244_create_currencies_table.php
@@ -16,7 +16,7 @@ class CreateCurrenciesTable extends Migration
         Schema::create('currencies', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name');
-            $table->double('currency');
+            $table->float('currency');
             $table->timestamps();
         });
     }
